Tsunami Beta ready & minor Sumo robot code changes
The title is fairly self-descriptive. I changed the Tsunami code we
wrote together so that things work better and I am finally getting
around to committing some Sumo robot code changes that I made a day
before the second competition.

+++ Vlad/TemperatureConversion_Output.cpp
@@ -28,11 +28,11 @@ int main ()
 	ofstream outfile(".\\Output\\converted_temperatures.txt");
 	
 	int number_of_rows = 0;
-	int file_content = 0;
+	double file_content = 0.0;
 	
 	while(infile_findLength >> file_content)
 	{
-	        number_of_rows++;
+		number_of_rows++;
 	}
 	       
 	//Declare variables pertaining to the 2D array containing all the temperature values
